% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Manipulate_table_function.R
\name{automat.abund.df}
\alias{automat.abund.df}
\title{Automatise the computation of abundance df for each Pose for a given
set of species}
\usage{
automat.abund.df(abund_allposes_list, species_set)
}
\arguments{
\item{abund_allposes_list}{a list containing the three dataframes of the
three poses}

\item{species_set}{a vector containing the name of species for which
dataframes must be retrieved}
}
\value{
the function returns a list for each species of list containing the
three df for the given species and the three poses
}
\description{
This function computes a list of abundance dataframe for the different
Poses for a given set of species or all species
}
