% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_SmaxN_across_camnb_functions.R
\name{clean.df.combcam.maxN}
\alias{clean.df.combcam.maxN}
\title{Create a dataframe containing SmaxN values for all studied species and
combinaison of cameras}
\usage{
clean.df.combcam.maxN(all_sp_list)
}
\arguments{
\item{all_sp_list}{a list containing for each species the output of the
compute.maxN.combcam() function. Names must be given for each
species. See script 4.SmaxN_increasing_nbcam_analysis.R part 3
"arrange data".}
}
\value{
a dataframe with six columns: species_nm, cam_nb, comb_nm, maxN, SmaxN
and SmaxN_timestep
}
\description{
This function computes a dataframes containing maxN,SmaxN, maxN_row,
for all species and the combination of cameras.
}
