% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_maxN_values_studied_species_functions.R
\name{clean.df.maxN}
\alias{clean.df.maxN}
\title{Create a dataframe containing maxN values for one studied species and
the three poses}
\usage{
clean.df.maxN(all_sp_list)
}
\arguments{
\item{all_sp_list}{a list containing for each species the output of the
automat.maxN.spbysp() function. Names must be given for each
species. See script 3.maxN_values_delta_analysis_spbysp.R part 2
"arrange data".}
}
\value{
a dataframe with five columns: species_nm, pose_nb, maxN, SmaxN
and maxN_timestep
}
\description{
This function computes a dataframes containing maxN,SmaxN, maxN_row,
for a one species and the three poses. This function uses
parallelisation.
}
