% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_SmaxN_across_camnb_functions.R
\name{combcam.plot}
\alias{combcam.plot}
\title{Create the plot showing SmaxN values for an increasing number of cameras}
\usage{
combcam.plot(maxN_combcam, colors, alpha, shape, size)
}
\arguments{
\item{maxN_combcam}{a dataframe containing all maxN values, deltas, species
names and poses (obtained from automat.maxN.setsp() function)}

\item{colors}{a vector containing the different colors for SmaxN and maxN}
}
\value{
a ggplot2 object showing the plots of SmaxN and maxN for each
species and this plot is saved as pdf in the output folder
}
\description{
This function computes plots showing SmaxN and maxN values across an
increasing number of cameras for 3h for each species (one plot = one sp)
}
