% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Duration_per_nbcam_functions.R
\name{compute.all.duration.nbcam}
\alias{compute.all.duration.nbcam}
\title{Create a dataframe containing number of seconds spend simulteaneously
in front on different camera number for ALL species}
\usage{
compute.all.duration.nbcam(abund_list)
}
\arguments{
\item{abund_list}{a list gathering the abundance dataframes of the different
Poses (dataframes with cameras = columns and time = rows)}
}
\value{
a dataframe with three columns: species name (same for all rows),
nb_cam (the number of cameras on which the species is seen simulteaneously)
and duration (the number of seconds the species is seen on a given number of
cameras)
}
\description{
This function computes for all species, a dataframe containing the number
of seconds the species is seen on different camera number simulteaneoulsy for
all the Poses
}
