% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Compute_combinaisons_functions.R
\name{compute.comb}
\alias{compute.comb}
\title{Create a list containing all combinaisons of n cameras}
\usage{
compute.comb(vect_cam)
}
\arguments{
\item{vect_cam}{a vector containing all camera names such as:
c("A1", "A2", "B1", "B2", "C1", "C2", "D", "E", "F", "G", "H", "I")}
}
\value{
the function returns a list with all combinaisons
}
\description{
This function computes a list which contains all combinaisons of n
cameras: all single cameras, then all pairs of cameras, then
all triplets of cameras etc with no repetition. This function is
based on:
https://yihui.org/cn/2007/04/generate-all-combinations-of-n-elements-in-r/
}
