% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Duration_per_nbcam_functions.R
\name{compute.duration.nbcam}
\alias{compute.duration.nbcam}
\title{Create a dataframe containing number of seconds spend simulteaneously
in front on different camera number for a given species}
\usage{
compute.duration.nbcam(species_nm, list_abund)
}
\arguments{
\item{species_nm}{the latin name of the studied species given with no
space but underscore "_" between Genera and Species names}

\item{list_abund}{a list gathering the abundance dataframes of the different
Poses (dataframes with cameras = columns and time = rows)}
}
\value{
a dataframe with three columns: species name (same for all rows),
nb_cam (the number of cameras on which the species is seen simulteaneously)
and duration (the number of seconds the species is seen on a given number of
cameras)
}
\description{
This function computes for a given species, a dataframe containing the number
of seconds the species is seen on different camera number simulteaneoulsy for
all the Poses
}
