% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_Stats_functions.R
\name{cor.SmaxN.plot}
\alias{cor.SmaxN.plot}
\title{Plot SmaxN vs maxN for each pose and do a correlation}
\usage{
cor.SmaxN.plot(maxN_all, color_poses, shape_sp, comp_metric)
}
\arguments{
\item{maxN_all}{a dataframe from the automat.maxN.setsp() function with SmaxN,
maxN, SmaxN_row, species names and Poses number columns}

\item{comp_metric}{the name of the metric to compare with the SmaxN. It could
either be "SmaxN_row" or "maxN". ONLY for the plot, correlations tests are
made for SmaxN_row and maxN.}

\item{colors_sp}{a vector containing the names of the colors to use to plot
the different species}

\item{shape_poses}{three numerical value referring to the shape of the points used to
plot SmaxN for each pose}
}
\value{
a plot saved in the output folder and the resultats of the correlation
test
}
\description{
This function computes a list of abundance dataframes for each species.
For each species, the dataframes contain the
full number of camers (ie 12)
}
