% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_SmaxN_across_camnb_functions.R
\name{create.abund.list.camcombn}
\alias{create.abund.list.camcombn}
\title{Create a list containing for each species, the abundance dataframes for all
cameras combinations}
\usage{
create.abund.list.camcombn(cam_set, abund_allcam_list)
}
\arguments{
\item{cam_set}{a vector containing the names of cameras to use (camera names
should be written as caracter stings)}

\item{abund_all_cam_list}{a list gathering the abundance of each species in a
given dataframes with the 12 cameras. One dataframe for one Pose. Retrieved from the
create.abundlist.allcam.poses() function.}
}
\value{
a list containing for each species, dataframes corresponding to the
combination of the different cameras
}
\description{
This function computes a list which contains for each species a list of
abundance dataframes with more or less columns (ie cameras) given the
combination of cameras it has been build for
}
