% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_SmaxN_time_functions.R
\name{create.abund.list.timespan}
\alias{create.abund.list.timespan}
\title{Create a list containing for each species, the abundance dataframes for all
tiemspans from 10min to 1h}
\usage{
create.abund.list.timespan(spans_set, abund_allcam_list)
}
\arguments{
\item{spans_set}{a vector containing the timespans to use (numeric values
refering to the number of seconds in the timespan. For ex: 10 minutes
timespan = 10*60sec = 600 seconds)}

\item{abund_cam_allposes_list}{a list gathering the abundance of each species in a
given dataframes with all poses fusioned and 12 cameras. Retrieved from the
create.abundlist.allcam.poses() function.}
}
\value{
a list containing for each species, dataframes corresponding to the
different timespans
}
\description{
This function computes a list which contains for each species a list of
abundance dataframes with more or less rows (ie time) given the
timespan it has been build for
}
