% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_SmaxN_across_camnb_functions.R
\name{create.abundlist.allcam.poses}
\alias{create.abundlist.allcam.poses}
\title{Create list of abundance dataframes with all cameras}
\usage{
create.abundlist.allcam.poses(cam_set, species_nm, abund_list)
}
\arguments{
\item{cam_set}{a vector containing the names of cameras to use (camera names
should be written as caracter stings)}

\item{species_nm}{a character string containing the name species to study (latin
names with no spaces and underscores "_" between Genera and species names)}

\item{abund_list}{a list gathering the abundance dataframes (one for each
pose) and the whole number of cameras}
}
\value{
a list with one dataframe per species
}
\description{
This function computes a list of abundance dataframes for each species.
For each species, the dataframes contain the
full number of camers (ie 12)
}
