% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_maxN_values_studied_species_functions.R
\name{deltas.plot}
\alias{deltas.plot}
\title{Create the plot showing delta between maxN values for a set of species}
\usage{
deltas.plot(maxN_all, colors)
}
\arguments{
\item{maxN_all}{a dataframe containing all maxN values, deltas, species
names and poses (obtained from automat.maxN.setsp() function)}

\item{colors}{a vector containiing the different colors for the three poses}
}
\value{
a ggplot2 plot showing delta1 = SmaxN - maxN and
delta2 = SmaxN - SmaxN_row (values on y) for a set of species (x) and the
different poses and different facets for deltas
}
\description{
This function computes a plot showing SmaxN, maxN, SmaxN_row (values on y)
for a given set of species (x) and the three poses (several colors)
and save the plot in outputs folder
}
