% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_SmaxN_across_camnb_functions.R
\name{final.combcam}
\alias{final.combcam}
\title{Create a dataframe containing maxN values for the studied species and
the different combinaison of cameras and the different poses}
\usage{
final.combcam(abund_combcam_list, dist_df, fish_speed, analysis_type)
}
\arguments{
\item{dist_df}{a numerical dataframe containing the distance between each
pair of camera. There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. \strong{Rows names and columns names
must be cameras names}. \strong{BE CAREFUL that the cameras are
the same and in the same order in the dist_df and in the abund_df!}}

\item{fish_speed}{a numerical value refering to the maximal speed of the
studied species. \strong{Speed must be given in meters per second}. If the
computation of maxN values should not take into account fish speed (that is
to say if the camera pooling is done at the second level),
\code{fish_speed = NULL}}

\item{analysis_type}{type of analysis to do
as this function can be used either for the combination of camera analysis
or for the timespan analysis. Can be either "combcam" for
the analysis of camera combination or "timespan" for the
analysis of timespans.}

\item{abund_poses_combcam_list}{a list gathering for all poses, a list for each species of
abundance dataframes for each combination of cameras (298 combinations).
Retrieved from the create.abund.list.camcombn() function. and looks like:
pose1 > pose1_sp1 > pose1_sp1_comb1 , pose1 > pose1_sp1 > pose1_sp1_comb2,
... , pose1 > pose1_sp2 > pose1_sp2_comb1, ...,
... , pose2 > pose2_sp1 > pose2_sp1_comb1 etc.}
}
\value{
a dataframe containing for each combination of cameras, maxN values
}
\description{
This function computes a dataframes containing maxN,SmaxN, maxN_row,
for a given set of species, all the combinaisons of a given set of
cameras and the different poses.
}
