% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Manipulate_table_function.R
\name{gather.abund.df}
\alias{gather.abund.df}
\title{Create a dataframe with abundance data for different cameras and a given
species}
\usage{
gather.abund.df(list_abund_df, species_nm, time_start, time_stop)
}
\arguments{
\item{list_abund_df}{a list containing as many dataframes as there are
cameras. Each dataframe is a species*abundance dataframe and contains the
abundance of different species (columns) across continuous time (rows)}

\item{species_nm}{a character string refering to the name of the species
for which dataframes must be gathered. \strong{Be sure that species names
is correctly spelled as in the vect_abund_df dataframes}.}

\item{time_start}{a character string refering to the time when the dataframe
should begin.}

\item{time_stop}{a character string refering to the time when the dataframe
should stop.}
}
\value{
the function returns a dataframe containing the abundance of a given
species for a given period of time. There are as many columns as there are
cameras and there are as many rows as there are timesteps (second) between
time_start and time_stop.
}
\description{
This function computes a dataframe containing the abundance of a given
species for a given period of time. The user gives as many dataframes as
there are cameras, a period of time and a species name and the function
checks on which cameras the species has been seen and gathers the dataframes
of these cameras.
}
