% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_Stats_functions.R
\name{kruskal.SmaxN.plot}
\alias{kruskal.SmaxN.plot}
\title{Plot SmaxN  across number of cameras or time and compute Kruskall Wallis test}
\usage{
kruskal.SmaxN.plot(SmaxN_df, metric)
}
\arguments{
\item{SmaxN_df}{a dataframe from the final.combcam() function with SmaxN
with an increasing number of cameras or with increasing recording time}

\item{metric}{a caracter string refering to the metric to which SmaxN must
be confronted. It can either be "cam_nb" or "timespan".}
}
\value{
a plot saved in the output folder and the results of the Kruskall-Wallis
test
}
\description{
This function computes Kruskall Wallis test to see if SmaxN is significantly
different across an increasing number of cameras or time
}
