% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Duration_per_nbcam_functions.R
\name{plot.duration.nbcam}
\alias{plot.duration.nbcam}
\title{Create plot showing the number of seconds spend simulteaneously
in front on different camera number for the 17 species}
\usage{
\method{plot}{duration.nbcam}(duration_cam_sp_df)
}
\arguments{
\item{duration_cam_sp_df}{a dataframe containing species, nb_cam and
duration columns}
}
\value{
a plot which is saved in the output folder
}
\description{
This function computes for all species, a plot: x = species,
y = duration of presence, colour = nb of cameras
}
