% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Plot_maxN_functions.R
\name{plot.maxN}
\alias{plot.maxN}
\title{Plot the maxN values}
\usage{
\method{plot}{maxN}(maxN_plot_df, color_points, alpha_points, color_boxplots, alpha_boxplots)
}
\arguments{
\item{maxN_plot_df}{a dataframe containing maxN values. This dataframe is
the ouput of the \code{prepare.maxN.plot() function}.}

\item{color_points}{a R color code for plotting points}

\item{alpha_points}{a transparency code for plotting points}

\item{color_boxplot}{a R color code for plotting boxplot(s)}

\item{alpha_boxplot}{a transparency code for plotting boxplot(s)}
}
\value{
the function returns a plot of maxN values
}
\description{
This function computes a plot containing the maxN values for different
definition of maxN. The user gives a data.frame which is the output of the
\code{prepare.maxN.plot() function}. The function returns the plot.
}
