% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Plot_maxN_functions.R
\name{prepare.maxN.plot}
\alias{prepare.maxN.plot}
\title{Create a dataframe with maxN data to plot for a given species}
\usage{
prepare.maxN.plot(list_maxN, indices_nm)
}
\arguments{
\item{list_maxN}{a list containing maxN values. This list is the ouput of
the \code{compute.max.abund() function} of the \code{SmaxN package}.}

\item{indices_nm}{a vector containing the names of maxN inidces to plot}
}
\value{
the function returns a dataframe values to plot
}
\description{
This function computes a dataframe containing the maxN values for different
definition of maxN. The user gives a
list of maxN values which is the output of the
\code{compute.max.abund() function} of the \code{SmaxN package}.
The function returns a df to be used to plot the maxN values.
}
