% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_maxN_values_studied_species_functions.R
\name{SmaxN.single.inp}
\alias{SmaxN.single.inp}
\title{Computes SmaxN and other abundance metrics using a list which have
the abundance dataframe, the speed and the distance dataframe as input}
\usage{
SmaxN.single.inp(i)
}
\arguments{
\item{list_input}{is a list which contains the abundance dataframe of the
studied species, the distance dataframe and the speed for a given pose}
}
\value{
the result of the \code{SmaxN.computation} function
}
\description{
Computes SmaxN and other abundance metrics using a list which have
the abundance dataframe, the speed and the distance dataframe as input
}
