% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_SmaxN_time_functions.R
\name{timespans.plot}
\alias{timespans.plot}
\title{Create the plot showing SmaxN values for an increasing amount of time}
\usage{
timespans.plot(maxN_timespans, colors, alpha, shape_pose, size)
}
\arguments{
\item{maxN_timespans}{a dataframe containing all maxN values, species
names, pose_nb, amount of time from the compute.maxN.combcam(analysis_type ="timespan")
function.}

\item{colors}{a vector containing the different colors for SmaxN and maxN}
}
\value{
a ggplot2 object showing the plots of SmaxN and maxN for each
species and this plot is saved as pdf in the output folder
}
\description{
This function computes plots showing SmaxN and maxN values across an
increasing amount of time for 12 cameras for each species (one plot = one sp)
}
